Add files via upload
Compress JPEG files using Pillow. Pillow is python library used for manipulation of images. The extent of the compression is dependant on the 'quality' parameter which you can alter depending on your need. Just copy the files into your images folder that need to compressed and run the script. A new folder named 'compressedImages' will be created. You will notice a line of code for PNG files. This is included only to handle images of PNG format if they happen to present in your source folder. PNG being a lossless format will not be effected by this script.